React router and redux data have been added.
The router has a single route implemented to make sure everything is connected.
In order to make things work nicely, the way of linking routes and components
needs to be thought through.

Redux is connecting to the user component. This was done with a simple reducer
with data being typed in manually. In the future, the data will be driven from
some sort of more permanent data store.