CSS styling updates and Webpack loading process
User table has some improved styling to match the design more closely. For now,
it has been mocked up in CSS. It will be converted to Sass if time permits.

The loading process now allows for CSS to integrate more tightly into React.
This will allow for cool things in the future like build specific CSS and assets
to allow for unique deployments.