Revamped route handling, added styling to user table, added placeholder for add user form.
The route handling now only updates the UI in the form bar section. The section
was re-factored from a component to a simple div to make things a bit easier.
The bar_form file has been deleted.

A style readedbg was added to the primary keys on the UI to allow for the circle
background. The initial code does not render the background, which needs to be
troubleshooted.

The user form is now present in the UI and a user can transition to it. This is
a placeholder for now and the full functionality needs to be implemented.